+++ amethyst_renderer/src/pass/sprite/interleaved.rs
@@ -1,48 +1,35 @@
-//! Simple flat forward drawing pass.
-
-use std::marker::PhantomData;
+//! Flat forward drawing pass that mimics a blit.
 
 use amethyst_assets::AssetStorage;
-use amethyst_core::specs::prelude::{Join, Read, ReadExpect, ReadStorage};
+use amethyst_core::specs::prelude::{Join, Read, ReadStorage};
 use amethyst_core::transform::GlobalTransform;
-use gfx::pso::buffer::ElemStride;
 use gfx_core::state::{Blend, ColorMask};
 use glsl_layout::Uniform;
 
 use super::*;
 use cam::{ActiveCamera, Camera};
 use error::Result;
-use mesh::{Mesh, MeshHandle};
-use mtl::{Material, MaterialDefaults};
-use pass::util::{draw_mesh, get_camera, setup_textures, VertexArgs};
+use mtl::MaterialTextureSet;
+use pass::util::{draw_sprite, get_camera, setup_textures, SpriteArgs, VertexArgs};
 use pipe::pass::{Pass, PassData};
 use pipe::{DepthMode, Effect, NewEffect};
+use sprite::{SpriteRender, SpriteSheet};
 use tex::Texture;
 use types::{Encoder, Factory};
-use vertex::{Position, Query, TexCoord};
-use visibility::Visibility;
+use sprite_visibility::SpriteVisibility;
 
-/// Draw mesh without lighting
-///
-/// See the [crate level documentation](index.html) for information about interleaved and separate
-/// passes.
-///
-/// # Type Parameters
-///
-/// * `V`: `VertexFormat`
+/// Draws sprites on a 2D quad.
 #[derive(Derivative, Clone, Debug, PartialEq)]
-#[derivative(Default(bound = "V: Query<(Position, TexCoord)>, Self: Pass"))]
-pub struct DrawFlat<V> {
-    _pd: PhantomData<V>,
+#[derivative(Default(bound = "Self: Pass"))]
+pub struct DrawSprite {
     transparency: Option<(ColorMask, Blend, Option<DepthMode>)>,
 }
 
-impl<V> DrawFlat<V>
+impl DrawSprite
 where
-    V: Query<(Position, TexCoord)>,
     Self: Pass,
 {
-    /// Create instance of `DrawFlat` pass
+    /// Create instance of `DrawSprite` pass
     pub fn new() -> Self {
         Default::default()
     }
@@ -59,27 +46,20 @@ where
     }
 }
 
-impl<'a, V> PassData<'a> for DrawFlat<V>
-where
-    V: Query<(Position, TexCoord)>,
-{
+impl<'a> PassData<'a> for DrawSprite {
     type Data = (
         Option<Read<'a, ActiveCamera>>,
         ReadStorage<'a, Camera>,
-        Read<'a, AssetStorage<Mesh>>,
+        Read<'a, AssetStorage<SpriteSheet>>,
         Read<'a, AssetStorage<Texture>>,
-        ReadExpect<'a, MaterialDefaults>,
-        Option<Read<'a, Visibility>>,
-        ReadStorage<'a, MeshHandle>,
-        ReadStorage<'a, Material>,
+        Read<'a, MaterialTextureSet>,
+        Option<Read<'a, SpriteVisibility>>,
+        ReadStorage<'a, SpriteRender>,
         ReadStorage<'a, GlobalTransform>,
     );
 }
 
-impl<V> Pass for DrawFlat<V>
-where
-    V: Query<(Position, TexCoord)>,
-{
+impl Pass for DrawSprite {
     fn compile(&mut self, effect: NewEffect) -> Result<Effect> {
         use std::mem;
         let mut builder = effect.simple(VERT_SRC, FRAG_SRC);
@@ -89,7 +69,11 @@ where
                 mem::size_of::<<VertexArgs as Uniform>::Std140>(),
                 1,
             )
-            .with_raw_vertex_buffer(V::QUERIED_ATTRIBUTES, V::size() as ElemStride, 0);
+            .with_raw_constant_buffer(
+                "SpriteArgs",
+                mem::size_of::<<SpriteArgs as Uniform>::Std140>(),
+                1,
+            );
         setup_textures(&mut builder, &TEXTURES);
         match self.transparency {
             Some((mask, blend, depth)) => builder.with_blended_output("color", mask, blend, depth),
@@ -106,69 +90,56 @@ where
         (
             active,
             camera,
-            mesh_storage,
+            sprite_sheet_storage,
             tex_storage,
-            material_defaults,
+            material_texture_set,
             visibility,
-            mesh,
-            material,
+            sprite_render,
             global,
         ): <Self as PassData<'a>>::Data,
     ) {
         let camera = get_camera(active, &camera, &global);
 
         match visibility {
-            None => for (mesh, material, global) in (&mesh, &material, &global).join() {
-                draw_mesh(
+            None => for (sprite_render, global) in (&sprite_render, &global).join() {
+                draw_sprite(
                     encoder,
                     effect,
-                    false,
-                    mesh_storage.get(mesh),
-                    None,
+                    sprite_render,
+                    &sprite_sheet_storage,
                     &tex_storage,
-                    Some(material),
-                    &material_defaults,
+                    &material_texture_set,
                     camera,
                     Some(global),
-                    &[V::QUERIED_ATTRIBUTES],
-                    &TEXTURES,
                 );
             },
             Some(ref visibility) => {
-                for (mesh, material, global, _) in
-                    (&mesh, &material, &global, &visibility.visible_unordered).join()
+                for (sprite_render, global, _) in
+                    (&sprite_render, &global, &visibility.visible_unordered).join()
                 {
-                    draw_mesh(
+                    draw_sprite(
                         encoder,
                         effect,
-                        false,
-                        mesh_storage.get(mesh),
-                        None,
+                        sprite_render,
+                        &sprite_sheet_storage,
                         &tex_storage,
-                        Some(material),
-                        &material_defaults,
+                        &material_texture_set,
                         camera,
                         Some(global),
-                        &[V::QUERIED_ATTRIBUTES],
-                        &TEXTURES,
                     );
                 }
 
                 for entity in &visibility.visible_ordered {
-                    if let Some(mesh) = mesh.get(*entity) {
-                        draw_mesh(
+                    if let Some(sprite_render) = sprite_render.get(*entity) {
+                        draw_sprite(
                             encoder,
                             effect,
-                            false,
-                            mesh_storage.get(mesh),
-                            None,
+                            sprite_render,
+                            &sprite_sheet_storage,
                             &tex_storage,
-                            material.get(*entity),
-                            &material_defaults,
+                            &material_texture_set,
                             camera,
                             global.get(*entity),
-                            &[V::QUERIED_ATTRIBUTES],
-                            &TEXTURES,
                         );
                     }
                 }