+++ examples/pong/main.rs
@@ -15,8 +15,8 @@ use amethyst::core::transform::TransformBundle;
 use amethyst::ecs::prelude::{Component, DenseVecStorage};
 use amethyst::input::InputBundle;
 use amethyst::prelude::*;
-use amethyst::renderer::{DrawFlat, PosTex};
-use amethyst::ui::UiBundle;
+use amethyst::renderer::{DisplayConfig, DrawSprite, Pipeline, RenderBundle, Stage};
+use amethyst::ui::{DrawUi, UiBundle};
 
 use audio::Music;
 use bundle::PongBundle;
@@ -49,6 +49,14 @@ fn main() -> amethyst::Result<()> {
         "{}/examples/pong/resources/display.ron",
         env!("CARGO_MANIFEST_DIR")
     );
+    let config = DisplayConfig::load(&display_config_path);
+
+    let pipe = Pipeline::build().with_stage(
+        Stage::with_backbuffer()
+            .clear_target([0.0, 0.0, 0.0, 1.0], 1.0)
+            .with_pass(DrawSprite::new())
+            .with_pass(DrawUi::new()),
+    );
 
     let key_bindings_path = {
         if cfg!(feature = "sdl_controller") {
@@ -68,19 +76,17 @@ fn main() -> amethyst::Result<()> {
 
     let game_data = GameDataBuilder::default()
         .with_bundle(
-            InputBundle::<String, String>::new().with_bindings_from_file(&key_bindings_path)?
-        )?
-        .with_bundle(PongBundle)?
+            InputBundle::<String, String>::new().with_bindings_from_file(&key_bindings_path)?,
+        )?.with_bundle(PongBundle)?
+        .with_bundle(RenderBundle::new(pipe, Some(config)).with_sprite_sheet_processor())?
         .with_bundle(TransformBundle::new().with_dep(&["ball_system", "paddle_system"]))?
         .with_bundle(AudioBundle::new(|music: &mut Music| music.music.next()))?
-        .with_bundle(UiBundle::<String, String>::new())?
-        .with_basic_renderer(display_config_path, DrawFlat::<PosTex>::new(), true)?;
+        .with_bundle(UiBundle::<String, String>::new())?;
     let mut game = Application::build(assets_dir, Pong)?
         .with_frame_limit(
             FrameRateLimitStrategy::SleepAndYield(Duration::from_millis(2)),
             144,
-        )
-        .build(game_data)?;
+        ).build(game_data)?;
     game.run();
     Ok(())
 }